TopNavBarDropwdownIconBug

+++ database/migrations/2025_03_12_075807_add_archived_to_notifications_table.php
@@ -8,15 +8,19 @@ class AddArchivedToNotificationsTable extends Migration
 {
     public function up()
     {
-        Schema::table('notifications', function (Blueprint $table) {
-            $table->boolean('archived')->default(false)->after('status');
-        });
+        if (!Schema::hasColumn('notifications', 'archived')) {
+            Schema::table('notifications', function (Blueprint $table) {
+                $table->boolean('archived')->default(0)->after('status');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('notifications', function (Blueprint $table) {
-            $table->dropColumn('archived');
-        });
+        if (Schema::hasColumn('notifications', 'archived')) {
+            Schema::table('notifications', function (Blueprint $table) {
+                $table->dropColumn('archived');
+            });
+        }
     }
-}   
+}
